feat: 토큰관련테스트 리팩토링
